requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<5
sqlalchemy==2.0.45
slowapi==0.1.9
limits>=2.3
python-dotenv==1.0.1
polars==1.8.2
pytest==8.3.4
httpx==0.27.2
psycopg2-binary==2.9.11
brutils==2.2.0